from php calling java service, need to conform to some conventions

+++ backend-www/app/Service/ProductServiceInterface.php
@@ -6,6 +6,6 @@ namespace App\Service;
 
 interface ProductServiceInterface
 {
-    public function add(string $name, double $price): string;
+    public function add(string $name, float $price): string;
 }
